requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.0
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
toml==0.10.1
trafaret==2.1.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
